+++ SpringProntoClin/src/main/java/com/application/SpringProntoClin/model/Paciente.java
@@ -17,7 +17,7 @@ public class Paciente {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idPaciente;
+    private Long idPaciente;
     private String nomePaciente;
     private String nomeSocial;
     private String telefonePaciente;
